pyproject.toml
[project]
name = "minddash-rag-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.123.5",
    "google-cloud-run>=0.12.0",
    "google-cloud-storage>=3.6.0",
    "langchain>=1.1.0",
    "langchain-google-vertexai>=3.1.1",
    "langchain-postgres>=0.0.16",
    "sqlmodel>=0.0.27",
]

[dependency-groups]
dev = [
    "ruff>=0.14.8",
]
